Clean up code for benchmarks and sampling strategies
